+++ src/components/buttons/CheckboxButton/CheckboxButton.tsx
@@ -2,10 +2,10 @@
 import { FC, ChangeEvent, FocusEvent } from 'react'
 import { jsx } from '@emotion/core'
 
-import { checkboxStyles, labelStyles, inputStyles, labelStylesDisabled } from './styles'
+import { checkboxStyles, labelStyles, inputStyles, labelStylesDisabled, childrenStyles, checkboxStylesTopSticked, checkboxStylesBottomSticked } from './styles'
 
 export const CheckboxButton: FC<CheckboxButtonProps> = props => (
-  <label css={[labelStyles, props.isDisabled ? labelStylesDisabled : null]}>
+  <label css={[labelStyles, props.isDisabled && labelStylesDisabled]}>
     <input
       css={[inputStyles]}
       type="checkbox"
@@ -16,8 +16,10 @@ export const CheckboxButton: FC<CheckboxButtonProps> = props => (
       onFocus={props.onFocus}
       onBlur={props.onBlur}
     />
-    <div css={[checkboxStyles]} />
-    {props.children}
+    <div css={[checkboxStyles, props.checkboxPosition === 'top' && checkboxStylesTopSticked, props.checkboxPosition === 'bottom' && checkboxStylesBottomSticked]} />
+    <div css={childrenStyles}>
+      {props.children}
+    </div>
   </label>
 )
 
@@ -25,6 +27,7 @@ export interface CheckboxButtonProps {
   value: string
   isDisabled?: boolean
   isChecked?: boolean
+  checkboxPosition?: 'top' | 'bottom' | 'center'
   onValueChange?: (event: ChangeEvent<HTMLInputElement>) => void
   onBlur?: (event: FocusEvent) => void
   onFocus?: (event: FocusEvent) => void